Move fetchers to app shared library
