+++ app/src/main/java/io/reark/rxgithubapp/advanced/data/DataStoreModule.java
@@ -39,7 +39,7 @@ import io.reark.rxgithubapp.advanced.data.stores.StoreModule;
 import io.reark.rxgithubapp.advanced.data.stores.UserSettingsStore;
 import io.reark.rxgithubapp.shared.injections.ForApplication;
 import io.reark.rxgithubapp.advanced.network.ServiceDataLayer;
-import io.reark.rxgithubapp.advanced.network.fetchers.FetcherModule;
+import io.reark.rxgithubapp.advanced.network.FetcherModule;
 
 @Module(includes = { FetcherModule.class, StoreModule.class })
 public final class DataStoreModule {
