+++ app/src/main/java/io/reark/rxgithubapp/advanced/data/stores/StoreBase.java
@@ -32,8 +32,10 @@ import com.google.gson.Gson;
 
 import io.reark.reark.data.store.SingleItemContentProviderStore;
 import io.reark.reark.utils.Preconditions;
+import io.reark.rxgithubapp.shared.data.StoreInterface;
 
-public abstract class StoreBase<T, U> extends SingleItemContentProviderStore<T, U> {
+public abstract class StoreBase<T, U> extends SingleItemContentProviderStore<T, U>
+        implements StoreInterface<T, U> {
 
     private final Gson gson;
