+++ appshared/src/main/java/io/reark/rxgithubapp/shared/data/StoreInterface.java
@@ -1,0 +1,14 @@
+package io.reark.rxgithubapp.shared.data;
+
+import android.net.Uri;
+
+import io.reark.rxgithubapp.shared.pojo.GitHubRepository;
+import io.reark.rxgithubapp.shared.pojo.GitHubRepositorySearch;
+
+/**
+ * Created by ttuo on 26/06/16.
+ */
+public interface StoreInterface <T, U> {
+    Uri getUriForId(U id);
+    void put(T item);
+}
